requirements.txt
celery==5.3.6
Django==4.2.7
djangorestframework==3.14.0
mysqlclient==2.2.0
redis==5.0.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0



